Simplify Scylla keyboard layout even further
Once I simplified my symbol layer by forcing myself to use the number
row for most symbols that freed up combining the number layer and the
symbol layer. Leaving me with a very compact 3 layer keyboard layout
that feels eficient to use and which is similar to the standard mac
keyboard layout.